+++ widgets/libs/extensions/swiper/swiper-common.scss
