+++ widgets/libs/extensions/swiper/swiper-font.scss
