+++ widgets/libs/extensions/swiper/swiper.extension.ts
@@ -1,6 +1,9 @@
 import Swiper from "swiper"
 import { SdkSwiper } from "types"
 import { Manipulation, Navigation } from "swiper/modules"
+import swiperFont from "./swiper-font.scss"
+import swiperCommon from "./swiper-common.scss"
+import swiperBundleCss from "@swiper/swiper-bundle.css"
 
 declare const sdk: SdkSwiper
 
@@ -23,6 +26,10 @@ export function initializeSwiper({
   nextButton = "swiper-button-next",
   initialIndex = 0
 }: SwiperProps) {
+  sdk.addWidgetCustomStyles(swiperFont)
+  sdk.addSharedCssCustomStyles(swiperBundleCss)
+  sdk.addSharedCssCustomStyles(swiperCommon)
+
   const prev = widgetSelector!.parentNode!.querySelector<HTMLElement>(`.${prevButton}`)
   const next = widgetSelector!.parentNode!.querySelector<HTMLElement>(`.${nextButton}`)
